:close #12 Entity User have been renamed to Client. Comes with test #16
