+++ config/doctrine/Client.orm.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                   https://raw.github.com/doctrine/doctrine2/master/doctrine-mapping.xsd">
+
+    <entity
+            name="App\Domain\Entity\User"
+            table="bm_user"
+            repository-class="App\Domain\Repository\UserRepository"
+    >
+        <id name="id" type="uuid" />
+        <field name="roles" type="array" />
+        <field name="createdAt" type="integer" />
+        <field name="updatedAt" type="integer" />
+        <field name="username" type="string" unique="true" />
+        <field name="password" type="string" />
+        <field name="phoneNumber" type="string" />
+        <field name="email" type="string" unique="true" />
+
+    </entity>
+
+</doctrine-mapping>
