+++ config/serializer/serializer_client.yaml
@@ -1,0 +1,10 @@
+App\Domain\Entity\User:
+    attributes:
+        username:
+            groups: ['users_list', 'user']
+        email:
+            groups: ['users_list', 'user']
+        phoneNumber:
+            groups: ['user']
+        createdAt:
+            groups: ['users_list']
