--- config/serializer/serializer_user.yaml
@@ -1,10 +1,0 @@
-App\Domain\Entity\User:
-    attributes:
-        username:
-            groups: ['users_list', 'user']
-        email:
-            groups: ['users_list', 'user']
-        phoneNumber:
-            groups: ['user']
-        createdAt:
-            groups: ['users_list']
