+++ src/Domain/DataFixtures/Fixtures/Client.yaml
@@ -1,0 +1,16 @@
+user_0:
+    Username: User0
+    Password: password0
+    Email: user0@mail.com
+user_1:
+    Username: User1
+    Password: password1
+    Email: user1@mail.com
+user_2:
+    Username: User2
+    Password: password2
+    Email: user2@mail.com
+user_3:
+    Username: RegularUser
+    Password: OpenClassrooms
+    Email: RegularUser@mail.com
