+++ src/UI/Action/Client/Interfaces/DeleteClientInterface.php
