--- src/UI/Action/User/Interfaces/ReadUserListInterface.php
@@ -1,34 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\UI\Action\User\Interfaces;
-
-use App\Domain\Repository\UserRepository;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Serializer\SerializerInterface;
-
-/**
- * Interface ReadUserListInterface
- * @package App\UI\Action\User\Interfaces
- */
-interface ReadUserListInterface
-{
-    /**
-     * ReadUserListInterface constructor.
-     *
-     * @param UserRepository $userRepository
-     * @param SerializerInterface $serializer
-     */
-    public function __construct(
-        UserRepository $userRepository,
-        SerializerInterface $serializer
-    );
-
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function __invoke(Request $request): Response;
-}
